+++ app/drawers/MyCharacters.vue
@@ -37,10 +37,9 @@ function handleDeleteCharacter(character: Character) {
   characterToDelete.value = character
 }
 
-function handleGoToChat(username: string) {
+async function handleGoToChat(username: string) {
   isRootDrawerOpen.value = false
-
-  navigateTo(`/app/chat/${username}`)
+  await navigateTo(`/app/chat/${username}`)
 }
 </script>