+++ db-client/username.sql
@@ -1,0 +1,9 @@
+-- ------------------ Username ------------------
+
+-- @block List all Usernames
+
+SELECT * FROM Username
+
+-- @block Update all Usernames.username to be lowercase
+
+UPDATE Username SET username = LOWER(username)