+++ db/sessions/contact.sql
@@ -1,7 +1,0 @@
--- @block List all contacts
-
-SELECT * FROM Contact
-
--- @block Delete all contacts
-
-DELETE FROM Contact